Update Kill Zones.lua

+++ SAPP SCRIPTS/Kill Zones.lua
@@ -15,9 +15,9 @@ https://github.com/Chalwk77/Halo-Scripts-Phasor-V2-/blob/master/LICENSE
 ]]--
 
 api_version = "1.11.0.0"
-coordiantes = { }
-kill_timer = { }
 players = { }
+kill_timer = { }
+coordiantes = { }
 warning_timer = { }
 kill_init_timer = { }
 -- ===================================================== CONFIGURATION STARTS ===================================================== --
@@ -33,18 +33,18 @@ kill_init_timer = { }
 
 --      label                      x,y,z                radius           Warning Dealy      Seconds until death
 coordiantes["bloodgulch"] = {
-    { "Kill Zone 1",      33.631, -65.569, 0.370,         5,                   0,                  15},
-    { "Kill Zone 2",      41.703, -128.663, 0.247,        5,                   0,                  15},
-    { "Kill Zone 3",      50.655, -87.787, 0.079,         5,                   0,                  15},
-    { "Kill Zone 4",      101.940, -170.440, 0.197,       5,                   0,                  15},
-    { "Kill Zone 5",      81.617, -116.049, 0.486,        5,                   0,                  15},
-    { "Kill Zone 6",      78.208, -152.914, 0.091,        5,                   0,                  15},
-    { "Kill Zone 7",      64.178, -176.802, 3.960,        5,                   0,                  15},
-    { "Kill Zone 8",      102.312, -144.626, 0.580,       5,                   0,                  15},
-    { "Kill Zone 9",      86.825, -172.542, 0.215,        5,                   0,                  15},
-    { "Kill Zone 10",     65.846, -70.301, 1.690,         5,                   0,                  15},
-    { "Kill Zone 11",     28.861, -90.757, 0.303,         5,                   0,                  15},
-    { "Kill Zone 12",     46.341, -64.700, 1.113,         5,                   0,                  15},
+    { "Kill Zone 1",      33.631, -65.569, 0.370,         1,                   0,                  15},
+    { "Kill Zone 2",      41.703, -128.663, 0.247,        1,                   0,                  15},
+    { "Kill Zone 3",      50.655, -87.787, 0.079,         1,                   0,                  15},
+    { "Kill Zone 4",      101.940, -170.440, 0.197,       1,                   0,                  15},
+    { "Kill Zone 5",      81.617, -116.049, 0.486,        1,                   0,                  15},
+    { "Kill Zone 6",      78.208, -152.914, 0.091,        1,                   0,                  15},
+    { "Kill Zone 7",      64.178, -176.802, 3.960,        1,                   0,                  15},
+    { "Kill Zone 8",      102.312, -144.626, 0.580,       1,                   0,                  15},
+    { "Kill Zone 9",      86.825, -172.542, 0.215,        1,                   0,                  15},
+    { "Kill Zone 10",     65.846, -70.301, 1.690,         1,                   0,                  15},
+    { "Kill Zone 11",     28.861, -90.757, 0.303,         1,                   0,                  15},
+    { "Kill Zone 12",     46.341, -64.700, 1.113,         1,                   0,                  15},
 }
 -- ===================================================== CONFIGURATION ENDS ======================================================= --
 function OnScriptLoad()
@@ -130,37 +130,39 @@ function OnTick()
                 -- validate coordiantes table
                 if coordiantes[mapname] ~= nil then
                     for j = 1, #coordiantes[mapname] do
-                        if coordiantes[mapname] ~= { } and coordiantes[mapname][j] ~= nil then
-                            -- check if player is in kill zone
-                            if inSphere(i, coordiantes[mapname][j][2], coordiantes[mapname][j][3], coordiantes[mapname][j][4], coordiantes[mapname][j][5]) == true then
-                                -- create new warning timer --
-                                players[get_var(i, "$n")].warning_timer = players[get_var(i, "$n")].warning_timer + 0.030
-                                -- monitor warning timer until it reaches the value of "Warning Dealy" (coordiantes[mapname][j][6])
-                                if players[get_var(i, "$n")].warning_timer >= math.floor(coordiantes[mapname][j][6]) then
-                                    -- clear the player's console to prevent duplicate messages (spam)
-                                    ClearConsole(i)
-                                    local minutes, seconds = secondsToTime(players[get_var(i, "$n")].warning_timer, 2)
-                                    -- initiate kill timer
-                                    kill_timer[i] = true
-                                    -- send player the warning
-                                    rprint(i, "Warning! You have entered " .. tostring(coordiantes[mapname][j][1]) .. ".")
-                                    rprint(i, "You will be killed in " .. coordiantes[mapname][j][7] - math.floor(seconds) .. " seconds if you don't leave this area!")
-                                end
-                                if (kill_timer[i] == true) then
-                                    -- create new kill timer
-                                    players[get_var(i, "$n")].kill_init_timer = players[get_var(i, "$n")].kill_init_timer + 0.030
-                                    -- monitor killer timer until it reaches the value of "Seconds until death" (coordiantes[mapname][j][7])
-                                    if players[get_var(i, "$n")].kill_init_timer >= math.floor(coordiantes[mapname][j][7]) then
+                        if player_alive(i) then
+                            if coordiantes[mapname] ~= { } and coordiantes[mapname][j] ~= nil then
+                                -- check if player is in kill zone
+                                if inSphere(i, coordiantes[mapname][j][2], coordiantes[mapname][j][3], coordiantes[mapname][j][4], coordiantes[mapname][j][5]) == true then
+                                    -- create new warning timer --
+                                    players[get_var(i, "$n")].warning_timer = players[get_var(i, "$n")].warning_timer + 0.030
+                                    -- monitor warning timer until it reaches the value of "Warning Dealy" (coordiantes[mapname][j][6])
+                                    if players[get_var(i, "$n")].warning_timer >= math.floor(coordiantes[mapname][j][6]) then
                                         -- clear the player's console to prevent duplicate messages (spam)
                                         ClearConsole(i)
-                                        -- reset timers --
-                                        kill_timer[i] = false
-                                        players[get_var(i, "$n")].warning_timer = 0
-                                        players[get_var(i, "$n")].kill_init_timer = 0
-                                        -- kill Player
-                                        execute_command("kill " ..i)
-                                        -- send player the unfateful message
-                                        rprint(i, "You were killed because you didn't leave " .. tostring(coordiantes[mapname][j][1]) .. " in time!")
+                                        local minutes, seconds = secondsToTime(players[get_var(i, "$n")].warning_timer, 2)
+                                        -- initiate kill timer
+                                        kill_timer[i] = true
+                                        -- send player the warning
+                                        rprint(i, "Warning! You have entered " .. tostring(coordiantes[mapname][j][1]) .. ".")
+                                        rprint(i, "You will be killed in " .. coordiantes[mapname][j][7] - math.floor(seconds) .. " seconds if you don't leave this area!")
+                                    end
+                                    if (kill_timer[i] == true) then
+                                        -- create new kill timer
+                                        players[get_var(i, "$n")].kill_init_timer = players[get_var(i, "$n")].kill_init_timer + 0.030
+                                        -- monitor killer timer until it reaches the value of "Seconds until death" (coordiantes[mapname][j][7])
+                                        if players[get_var(i, "$n")].kill_init_timer >= math.floor(coordiantes[mapname][j][7]) then
+                                            -- clear the player's console to prevent duplicate messages (spam)
+                                            ClearConsole(i)
+                                            -- reset timers --
+                                            kill_timer[i] = false
+                                            players[get_var(i, "$n")].warning_timer = 0
+                                            players[get_var(i, "$n")].kill_init_timer = 0
+                                            -- kill Player
+                                            execute_command("kill " ..i)
+                                            -- send player the unfateful message
+                                            rprint(i, "You were killed because you didn't leave " .. tostring(coordiantes[mapname][j][1]) .. " in time!")
+                                        end
                                     end
                                 end
                             end
